+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## CHANGES in 0.10:
+- Bug: use actual envelope header if present (#412)
 - Bug: stats.lua should use UTC for calculating year and month (#409)
 - Bug: atom.lua - should generate date in UTC, not local time (#408)
 - Bug: shorten links does not work with cluster ids (#406)